+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add context menu for chat messages to allow rerolls for valid d20 rolls from the chat pane.
+- Add One Reroll setting (this only works for context menu rerolls).
 - Add CHANGELOG.md.
 - Update Github Workflows.
 - Add message handler.